backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
# psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests>=2.31.0
pandas
numpy
# TA-Lib==0.4.28
TA-Lib
yfinance
scikit-learn
nltk
textblob
reportlab==4.0.8
jinja2==3.1.2
beautifulsoup4==4.12.2
celery==5.3.6
redis==5.0.1
# gunicorn==21.2.0
pytest==7.4.4
matplotlib
pyotp>=2.9.0
openai>=1.0.0
lightgbm
